refactor: Refactor data model directory
