+++ app/src/main/java/edu/skku/cs/dokkang/data_models/request/LoginRequest.java
@@ -1,8 +1,7 @@
-package edu.skku.cs.dokkang.data_models.response;
+package edu.skku.cs.dokkang.data_models.request;
 
-public class SignUpDataModel {
+public class LoginRequest {
     private String username;
-    private String email;
     private String password;
 
     public String getUsername() {
@@ -11,14 +10,6 @@ public class SignUpDataModel {
 
     public void setUsername(String username) {
         this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
     }
 
     public String getPassword() {
