+++ app/src/main/res/layout/activity_community_main.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CommunityMain">
+    tools:context=".activities.CommunityMain">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
