+++ app/src/main/res/layout/activity_personal_info.xml
@@ -5,6 +5,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PersonalInfo">
+    tools:context=".activities.PersonalInfo">
 
   </androidx.constraintlayout.widget.ConstraintLayout>
