+++ app/src/main/res/layout/activity_study_group_main.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudyGroupMain">
+    tools:context=".activities.StudyGroupMain">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
