+++ app/src/main/res/layout/activity_subject_edit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SubjectEdit">
+    tools:context=".activities.SubjectEdit">
 
     <Button
         android:id="@+id/confirmButton"
